export debug module
