+++ lib/codecept.js
@@ -72,7 +72,7 @@ class Codecept {
     if (!this.config.noGlobals) {
       global.Helper = global.codecept_helper = require('@codeceptjs/helper')
       global.actor = global.codecept_actor = require('./actor')
-      global.pause = require('./pause')
+      global.pause = require('./debug').pause
       global.within = require('./within')
       global.session = require('./session')
       global.DataTable = require('./data/table')
