+++ lib/command/interactive.js
@@ -43,7 +43,7 @@ module.exports = async function (path, options) {
         break
       }
     }
-    require('../pause')()
+    require('../debug').pause()
     // recorder.catchWithoutStop((err) => console.log(err.stack));
     recorder.add(() => event.emit(event.test.after, {}))
     recorder.add(() => event.emit(event.suite.after, {}))
