+++ lib/helper/AI.js
@@ -8,7 +8,7 @@ const Container = require('../container')
 const { splitByChunks, minifyHtml } = require('../html')
 const { beautify } = require('../utils')
 const output = require('../output')
-const { registerVariable } = require('../pause')
+const { registerVariable } = require('../debug')
 
 const gtpRole = {
   user: 'user',
