+++ lib/index.js
@@ -27,7 +27,7 @@ module.exports = {
   /** @type {typeof CodeceptJS.Helper} */
   Helper: require('./helper'),
   /** @type {typeof CodeceptJS.pause} */
-  pause: require('./pause'),
+  pause: require('./debug').pause,
   /** @type {typeof CodeceptJS.within} */
   within: require('./within'),
   /**  @type {typeof CodeceptJS.DataTable} */
@@ -43,4 +43,4 @@ module.exports = {
   ai: require('./ai'),
 
   Workers: require('./workers'),
-};
+}
