+++ lib/plugin/pauseOnFail.js
@@ -1,5 +1,5 @@
 const event = require('../event')
-const pause = require('../pause')
+const { pause } = require('../debug')
 
 /**
  * Automatically launches [interactive pause](/basics/#pause) when a test fails.
